Move SwingGUI to it's own package to reduce clutter in case of multiple GUIs and update References
