+++ src/com/musicplayer/gui/swing/Button.java
@@ -1,4 +1,4 @@
-package com.musicplayer.gui;
+package com.musicplayer.gui.swing;
 
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
