+++ src/com/musicplayer/gui/swing/CoverDisplayType.java
@@ -1,4 +1,4 @@
-package com.musicplayer.gui;
+package com.musicplayer.gui.swing;
 
 /**
  * Options for cover display
